Support auth with custom service principal name
Add an optional element to the hosts spec, AUTH_NAME, which is
the "canonical" name to use for the service principal name instead
of HOSTNAME (or its corresponding fqdn). This option is useful in case of
accessing an authenticated server via ssh port forwarding, in which case
the HOSTNAME is 127.0.0.1.